requirements.txt
requests
pycairo
PyGObject
websocket-client
obs-ws-rc
